Move anyForSubtypeOf() in jqwikGlobals
